requirements.txt
pandas>=2.0.0
kafka-python
python-dotenv
pyspark==3.5.4
kaggle
google-cloud-bigquery>=3.0.0
db-dtypes
pandas-gbq
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
psycopg2